Refactored the code - Added only one package for all modules - Changes for base Loader class - Changes for one logger for entire package
