+++ examples/rethink_example.py
@@ -6,8 +6,7 @@ import json
 import os
 import sys
 import logging
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), './dbloader')))
-import dbloader as dbl
+from dbloader import dbloader as dbl
 
 logger = logging.getLogger(__name__)
