+++ setup.py
@@ -15,11 +15,7 @@ setup(name = 'dbloader',
       author = 'Phil Hildebrand',
       author_email = 'phil.hildebrand@gmail.com',
       license = 'mit',
-      packages = ['dbloader', 'loader.loader', 'mongo.mongo_loader', 'rethink.rethink_loader'],
-      package_dir = {'dbloader': 'dbloader',
-                     'loader.loader': 'dbloader/loader',
-                     'mongo.mongo_loader' : 'dbloader/mongo',
-                     'rethink.rethink_loader' : 'dbloader/rethink'},
+      packages = ['dbloader'],
       install_requires = [ requirements ],
       test_suite="pytest",
       use_2to3=True
